Added platform-specific implementation Fixed compilation errors on unix (for x64 only, not tested on x86 yet) Improved support for Unix
Network/*:
- Using FunctionList instead of MOHPC_HANDLERLIST_xxx macros

Network/InfoTypes:
- Added a functions on usercmd_t to manipulate button flags
- Removed union flags fron usercmd_t

Tests:
- Moved Log from Network to Main for global use
- Fixed command input on Network test

+++ mohpc-lib/Network/Client/MasterList.cpp
@@ -15,7 +15,7 @@ using namespace Network;
 
 #define MOHPC_LOG_NAMESPACE "masterlist"
 
-static const uint8_t gameKeys[gameListType_e::max][7] =
+static const uint8_t gameKeys[(size_t)gameListType_e::max][7] =
 {
 	// MOH:AA
 	{ 'M', '5', 'F', 'd', 'w', 'c' },
@@ -25,7 +25,7 @@ static const uint8_t gameKeys[gameListType_e::max][7] =
 	{ 'y', '3', '2', 'F', 'D', 'c' }
 };
 
-static const char* gameName[gameListType_e::max] =
+static const char* gameName[(size_t)gameListType_e::max] =
 {
 	"mohaa",
 	"mohaas",
